move the EvolvingAnsatzMinimumEigensolverResult to a new file
